requirements.txt
requests>=2.31.0
click>=8.1.0
mysql-connector-python>=8.0.0
